Moved most of the PCA preprocessing and running into the analysis_utils.py file to declutter the notebook.
Added SOA_is_NA feature to PCA preprocessing as a way to deal with the
issue introduced by the overloading of SOA to denote not just the
asynchrony between two tasks' stimuli, but the asynchrony between a
target and distractor in interference tasks. I was using -9999 as a
placeholder for "N/A", but I saw that when I went back into the original
parameter space from the PC-space after interpolation, that I was
getting large negative SOAs from the skew of the "N/A" values. This
fixes that problem by introducing a new dimension. When it's 1, it means
we're dealing with univalent stimuli (except in the Allport et al
experiment 5 case). When it's 0, that means there's a meaningful SOA and
we could be working with interference or PRP.

Also fixed an bug where task 2 difficulty was getting recorded as -1
even when there was a valid difficulty there, in task switching cases.

Updated notes to reflect the overloading of the switch rate for handling
uncertain task order and the changes to SOA. Pre-emptively added a note
about SOA overrides but I haven't implemented this in the convert.py
script yet.

After all the bug fixes, the PCA provides a different set of principal
components and this is seen in the updated notebook. I also interpolated
between 1) dual-task and interference, 2) dual-task and task-switching,
and 3) task-switching and interference. Those results are in the
notebook.